Calculator calcul doubles

+++ RPNCalculator-Kata.Tests/RPNOperatorTest.cs
@@ -8,13 +8,13 @@ namespace RPNCalculator_Kata.Tests
     {
         [TestCase(3, 5, 8)]
         [TestCase(4, 5, 9)]
-        public void Sould_Sum_Two_Number(int x, int y, int expectedResult)
+        public void Sould_Sum_Two_Number(double x, double y, double expectedResult)
         {
             // GIVEN
-            Func<int, int, int> sum = RPNOperator.Sum;
+            Func<double, double, double> sum = RPNOperator.Sum;
 
             // WHEN
-            int result = sum(x, y);
+            double result = sum(x, y);
 
             // THEN
             Check.That(result).IsEqualTo(expectedResult);
@@ -22,13 +22,13 @@ namespace RPNCalculator_Kata.Tests
 
         [TestCase(3, 5, -2)]
         [TestCase(44, 5, 39)]
-        public void Sould_Substract_Two_Number(int x, int y, int expectedResult)
+        public void Sould_Substract_Two_Number(double x, double y, double expectedResult)
         {
             // GIVEN
-            Func<int, int, int> substract = RPNOperator.Substract;
+            Func<double, double, double> substract = RPNOperator.Substract;
 
             // WHEN
-            int result = substract(x, y);
+            double result = substract(x, y);
 
             // THEN
             Check.That(result).IsEqualTo(expectedResult);
@@ -36,13 +36,13 @@ namespace RPNCalculator_Kata.Tests
 
         [TestCase(30, 5, 6)]
         [TestCase(15, 3, 5)]
-        public void Sould_Divide_Two_Number(int x, int y, int expectedResult)
+        public void Sould_Divide_Two_Number(double x, double y, double expectedResult)
         {
             // GIVEN
-            Func<int, int, int> divide = RPNOperator.Divide;
+            Func<double, double, double> divide = RPNOperator.Divide;
 
             // WHEN
-            int result = divide(x, y);
+            double result = divide(x, y);
 
             // THEN
             Check.That(result).IsEqualTo(expectedResult);
@@ -50,13 +50,13 @@ namespace RPNCalculator_Kata.Tests
 
         [TestCase(3, 5, 15)]
         [TestCase(15, 3, 45)]
-        public void Sould_Multiply_Two_Number(int x, int y, int expectedResult)
+        public void Sould_Multiply_Two_Number(double x, double y, double expectedResult)
         {
             // GIVEN
-            Func<int, int, int> multiply = RPNOperator.Multiply;
+            Func<double, double, double> multiply = RPNOperator.Multiply;
 
             // WHEN
-            int result = multiply(x, y);
+            double result = multiply(x, y);
 
             // THEN
             Check.That(result).IsEqualTo(expectedResult);
