+++ RPNCalculator-Kata/RPNOperator.cs
@@ -2,22 +2,22 @@
 {
     public class RPNOperator
     {
-        public static int Sum(int x, int y)
+        public static double Sum(double x, double y)
         {
             return x + y;
         }
 
-        public static int Substract(int x, int y)
+        public static double Substract(double x, double y)
         {
             return x - y;
         }
 
-        public static int Divide(int x, int y)
+        public static double Divide(double x, double y)
         {
             return x / y;
         }
 
-        public static int Multiply(int x, int y)
+        public static double Multiply(double x, double y)
         {
             return x * y;
         }
